Moved common functionality to OoniCrawler baseclass, currently only implemented for facebook for testing.
